Refactor : MVC 패턴 구조화
