+++ src/main/java/christmas/Application.java
@@ -1,6 +1,7 @@
 package christmas;
 
 import camp.nextstep.edu.missionutils.Console;
+import christmas.controller.Planner;
 
 public class Application {
     public static void main(String[] args) {
