+++ src/main/java/christmas/util/Converter.java
@@ -1,7 +1,8 @@
-package christmas;
+package christmas.util;
 
 import christmas.exception.ErrorMessage;
 import christmas.exception.InvalidDateException;
+import christmas.global.Const;
 import java.util.Arrays;
 import java.util.Map;
 import java.util.stream.Collectors;
